0409 - enhance firefox install progress

+++ Set-ExecutionPolicy Bypass -f.ps1
@@ -37,8 +37,14 @@ catch {
     # Write-Output $Error[0]
 }
 
+$install = Read-Host "Install Firefox? (y/n)";
+if ( $install -eq "y" )
+{
 wget -O C:/FirefoxSetup.exe "https://download.mozilla.org/?product=firefox-latest&os=win&lang=en-US"
 Start-Process C:/FirefoxSetup.exe
+}
+elseif ( $install -eq "n") { }
+else { }
 
 rm -r 'C:\tools\*.bat'
 rm -r 'C:\Users\mcu\*.bat'
